get react-router app rendering properly again
